pyproject.toml
[tool.poetry]
name = "neural-fonts"
version = "1.2.0"
description = "Hangul Font Creation based on GAN"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
Pillow = "^9.1.1"
numpy = "^1.22.4"
scipy = "^1.8.1"
tensorflow = "^2.9.1"
imageio = "^2.19.3"
scikit-image = "^0.19.3"
opencv-python = "^4.6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-black = "^0.3.3"
flake8-type-checking = "^1.5.0"
flake8-annotations = "^2.9.0"
flake8-isort = "^4.1.1"
pyright = "^1.1.254"
pytest = "^7.1.2"
poethepoet = "^0.13.1"

[tool.poetry.scripts]
infer = "neural_fonts.infer:main"
train = "neural_fonts.train:main"
package = "neural_fonts.package:main"
crop = "neural_fonts.crop:main"
font2img = "neural_fonts.font2img:main"

[tool.poe.tasks]
format = "black neural_fonts"
"check:flake8" = "flake8 neural_fonts"
"check:pyright" = "pyright"

[tool.isort]
profile = "black"

[tool.pyright]
include = ["neural_fonts"]
strict = ["neural_fonts"]
reportMissingTypeStubs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
